# docs-website/static/docs/api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for the traduora REST API\n\nOfficial website: https://traduora.com\nAdditional documentation: https://docs.traduora.com\nSource code: https://github.com/traduora/traduora",
    "version": "0.14.0",
    "title": "Traduora API"
  },
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "password",
      "authorizationUrl": "/api/v1/auth/token",
      "tokenUrl": "/api/v1/auth/token"
    }
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/translations": {
      "get": {
        "summary": "List all translation locales for a project",
        "parameters": [
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ListProjectLocalesResponse"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Translations"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "post": {
        "summary": "Add a new translation locale for a project",
        "parameters": [
          {
            "name": "AddLocaleRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/AddLocaleRequest"
            }
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/ProjectLocaleResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Plan limit reached"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Translations"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
    "/api/v1/projects/{projectId}/translations/{localeCode}": {
      "get": {
        "summary": "List translated terms for a locale",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ListTermTranslatonsResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project not found"
          }
        },
        "tags": ["Translations"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "patch": {
        "summary": "Update a term's translation",
        "parameters": [
          {
            "name": "UpdateTranslationRequest",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/UpdateTranslationRequest"
            }
          },
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "schema": {
              "$ref": "#/definitions/TermTranslatonResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project or locale not found"
          }
        },
        "tags": ["Translations"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Delete a project's locale",
        "description": "Deletes a project's locale and all related translations",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project or locale not found"
          }
        },
        "tags": ["Translations"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
    "/api/v1/projects/{projectId}/labels/{labelId}/terms/{termId}/translations/{localeCode}": {
      "post": {
        "summary": "Label a project term translation",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      },
      "delete": {
        "summary": "Unlabel a project term translation",
        "parameters": [
          {
            "type": "string",
            "name": "localeCode",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "termId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "labelId",
            "required": true,
            "in": "path"
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Label removed"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Project, term or label not found"
          }
        },
        "tags": ["Project Labels"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"],
        "consumes": ["application/json"]
      }
    },
...
  },
  "definitions": {
...
    "AccessDatesDTO": {
      "type": "object",
      "properties": {
        "created": {
          "type": "string"
        },
        "modified": {
          "type": "string"
        }
      },
      "required": ["created", "modified"]
    },
...
    "LocaleDTO": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "language": {
          "type": "string"
        },
        "region": {
          "type": "string"
        }
      },
      "required": ["code", "language", "region"]
    },
    "ProjectLocaleDTO": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "locale": {
          "$ref": "#/definitions/LocaleDTO"
        },
        "date": {
          "$ref": "#/definitions/AccessDatesDTO"
        }
      },
      "required": ["id", "locale", "date"]
    },
    "ListProjectLocalesResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ProjectLocaleDTO"
          }
        }
      },
      "required": ["data"]
    },
    "AddLocaleRequest": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "minLength": 2,
          "maxLength": 16
        }
      },
      "required": ["code"]
    },
    "ProjectLocaleResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ProjectLocaleDTO"
        }
      },
      "required": ["data"]
    },
    "TermTranslationDTO": {
      "type": "object",
      "properties": {
        "termId": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "labels": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "date": {
          "$ref": "#/definitions/AccessDatesDTO"
        }
      },
      "required": ["termId", "value", "labels", "date"]
    },
    "ListTermTranslatonsResponse": {
      "type": "object",
      "properties": {
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TermTranslationDTO"
          }
        }
      },
      "required": ["data"]
    },
    "UpdateTranslationRequest": {
      "type": "object",
      "properties": {
        "termId": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "minLength": 0,
          "maxLength": 8192
        }
      },
      "required": ["termId", "value"]
    },
    "TermTranslatonResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/TermTranslationDTO"
        }
      },
      "required": ["data"]
    },
...
  }
}